Changed damage from double to int

+++ src/main/java/shared/gameObjects/weapons/Gun.java
@@ -36,7 +36,7 @@ abstract class Gun extends Weapon {
       double sizeX,
       double sizeY,
       ObjectID id,
-      double damage,
+      int damage,
       double weight,
       String name,
       int ammo,
